authenticateUser dumps password to output
When calling AuthenticationApi::authenticateUser() the data being posted to the API (which contains the username and password) is dumped to the script output. If the client happens to be called in-process from a web page then the API username and password gets dumped to the visitors page which can then be used to compromise the user's exavault account.

Check ApiClient.php line 218, recommend removing the "print_r()".

Also line 212 in the same file. I'm not sure that 'echo' needs to be there either.